Yu jin (#16)
* Add files via upload

* Delete mechanism for linux.sh

* script on windows

This script gets tools on windows and uses the CMD to finish some steps, but I don't how to run the .sh file on windows. So it just installs the files on windows.

* Update script.bat

* Update script.bat

* Update script.bat

It, which will save in a temp file, has been changed by Yu. After that, I also change something about setting the environment variable. If there is a question, please let me know.

* Update script.bat

The script will ask the user where they want to install them. After that, it will be not to use any command that the windows does not come with. In the end, the script will set the system environment variable.